Add files via upload
I am using the 'Imodel' equation (5.7) to obtain values for multiple parameters (phi, beta, Ioff, I0) by fitting a curve to my experimental data. 
I am then using those parameters to calculate a guess for f' (the real component of the anomalous scattering factor). 
Thirdly, I perform a Kramers-Kronig transform to generate a guess for f" (the imaginary component). 
These f' and f" are put back into the Imodel to get a better guess for the parameters.
The process is repeated until the parameter values stop improving, at which point the f' and f" are deemed to be correct.